Final project - Observing Proposal
This is a JWST NIRSpec proposal to obtain host spectra for SN 2023adsy (highest z SN with z~2.9). In an ideal world, I would have modified the Scolnic mass step plot (Fig 3) but the code is in python 2. For the submission of the actual proposal I will talk to the people who made the original and find out how to extend the plot to show the redshift this SN was found at. This is what I was able to manage at this point. 

Also any comments on this would be super helpful cause I really do want to submit this!! Thanks!